refactor: [006-GROUP/CREATE] entity폴더 추가로 인한 import 조정
