+++ src/main/java/com/valuewith/tweaver/chat/repository/ChatRoomRepository.java
@@ -1,6 +1,7 @@
 package com.valuewith.tweaver.chat.repository;
 
-import com.valuewith.tweaver.chat.ChatRoom;
+
+import com.valuewith.tweaver.chat.entity.ChatRoom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
