+++ src/main/java/com/valuewith/tweaver/chat/service/ChatRoomService.java
@@ -1,7 +1,8 @@
 package com.valuewith.tweaver.chat.service;
 
-import com.valuewith.tweaver.chat.ChatRoom;
-import com.valuewith.tweaver.group.Group;
+
+import com.valuewith.tweaver.chat.entity.ChatRoom;
+import com.valuewith.tweaver.group.entity.Group;
 
 public interface ChatRoomService {
   ChatRoom setChatRoom(Group group);
