+++ src/main/java/com/valuewith/tweaver/group/repository/GroupRepository.java
@@ -1,6 +1,6 @@
 package com.valuewith.tweaver.group.repository;
 
-import com.valuewith.tweaver.group.Group;
+import com.valuewith.tweaver.group.entity.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
