+++ src/main/java/com/valuewith/tweaver/group/service/GroupService.java
@@ -1,7 +1,7 @@
 package com.valuewith.tweaver.group.service;
 
-import com.valuewith.tweaver.group.Group;
 import com.valuewith.tweaver.group.dto.GroupDto;
+import com.valuewith.tweaver.group.entity.Group;
 
 public interface GroupService {
   Group writeGroup(GroupDto groupDto);
