+++ src/main/java/com/valuewith/tweaver/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.valuewith.tweaver.member.repository;
 
-import com.valuewith.tweaver.member.Member;
+import com.valuewith.tweaver.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
