+++ src/main/java/com/valuewith/tweaver/member/service/MemberService.java
@@ -1,8 +1,8 @@
 package com.valuewith.tweaver.member.service;
 
-import com.valuewith.tweaver.chat.ChatRoom;
-import com.valuewith.tweaver.group.Group;
-import com.valuewith.tweaver.user.User;
+import com.valuewith.tweaver.chat.entity.ChatRoom;
+import com.valuewith.tweaver.group.entity.Group;
+import com.valuewith.tweaver.user.entity.User;
 
 public interface MemberService {
   void setMember(Group group, User user, ChatRoom chatRoom);
