+++ src/main/java/com/valuewith/tweaver/place/repository/PlaceRepository.java
@@ -1,6 +1,6 @@
 package com.valuewith.tweaver.place.repository;
 
-import com.valuewith.tweaver.place.Place;
+import com.valuewith.tweaver.place.entity.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
