+++ src/main/java/com/valuewith/tweaver/place/service/PlaceService.java
@@ -1,6 +1,6 @@
 package com.valuewith.tweaver.place.service;
 
-import com.valuewith.tweaver.group.Group;
+import com.valuewith.tweaver.group.entity.Group;
 import com.valuewith.tweaver.place.dto.PlaceDto;
 import java.util.List;
